# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Pienvarastokekus REST API",
    "contact": {
      "email": "suihkonen.matias@gmail.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "Agreement",
      "description": ""
    }
  ],
  "paths": {
    "/agreement/list": {
      "get": {
        "tags": [
          "Agreement"
        ],
        "summary": "Get all agreements",
        "description": "",
        "operationId": "placeOrder",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Agreement"
            }
          },
          "400": {
            "description": "Invalid Order"
          }
        }
      }
    },
    "/agreement/create": {
      "post": {
        "tags": [
          "Agreement"
        ],
        "summary": "Create agreement",
        "description": "Create pienvarastokeskus agreement file for pdf format",
        "operationId": "getOrderById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userdata",
            "in": "body",
            "description": "object for userdata",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "address": {
                  "type": "string"
                },
                "phoneNumber": {
                  "type": "string"
                }
              }

            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Agreement.create"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    }
  },
  "definitions": {

    "Agreement": {
      "type": "object",
      "properties": {
        "file": {
          "type": "string"
        }
      }
    },

    "Agreement.create": {
      "type": "object",
      "properties": {
        "filename": {
          "type": "string"
        }
      }
    }
  }
}